% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miniimagematrix}
\alias{miniimagematrix}
\title{Mini-image data matrice for example}
\format{
A list with two grayscale image matrices
\describe{
\item{A}{A matrix for one image}
\item{B}{A matrix for a second image}
}
}
\usage{
data(miniimagematrix)
}
\description{
The matrices are grayscale images in matrix format.
}
\examples{
 \donttest{
data(miniimagematrix)
miniimagematrix$A
}
}
\keyword{datasets}
